Refactor: sidebarFooter  component
